Rename nc_olci_l1b/l2 readers to olci_l1b/l2
