+++ satpy/etc/readers/olci_l1b.yaml
@@ -1,23 +1,23 @@
 reader:
   description: NC Reader for OLCI data
-  name: nc_olci_l1b
+  name: olci_l1b
   sensors: [olci]
   default_channels: []
   reader: !!python/name:satpy.readers.yaml_reader.FileYAMLReader
 
 file_types:
     esa_l1b:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCI1B
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCI1B
         file_patterns: ['{mission_id:3s}_OL_1_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{platform_mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/{dataset_name}_radiance.nc']
         requires: [esa_cal]
     esa_angles:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCIAngles
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCIAngles
         file_patterns: ['{mission_id:3s}_OL_1_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{platform_mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/tie_geometries.nc']
     esa_geo:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCIGeo
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCIGeo
         file_patterns: ['{mission_id:3s}_OL_1_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{platform_mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/geo_coordinates.nc']
     esa_cal:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCICal
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCICal
         file_patterns: ['{mission_id:3s}_OL_1_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{platform_mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/instrument_data.nc']
 datasets:
   longitude:
