+++ satpy/etc/readers/olci_l2.yaml
@@ -1,6 +1,6 @@
 reader:
   description: NC Reader for OLCI data
-  name: nc_olci_l2
+  name: olci_l2
   sensors: [olci]
   default_channels: []
   reader: !!python/name:satpy.readers.yaml_reader.FileYAMLReader
@@ -8,22 +8,22 @@ reader:
 file_types:
 
     esa_l2_reflectance:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCI2
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCI2
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/{dataset_name}_reflectance.nc']
     esa_l2_chl_nn:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCI2
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCI2
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/chl_nn.nc']
     esa_l2_chl_oc4me:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCI2
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCI2
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/chl_oc4me.nc']
     esa_l2_wqsf:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCI2
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCI2
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/wqsf.nc']
     esa_angles:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCIAngles
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCIAngles
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/tie_geometries.nc']
     esa_geo:
-        file_reader: !!python/name:satpy.readers.nc_olci.NCOLCIGeo
+        file_reader: !!python/name:satpy.readers.olci_nc.NCOLCIGeo
         file_patterns: ['{mission_id:3s}_OL_2_{datatype_id:_<6s}_{start_time:%Y%m%dT%H%M%S}_{end_time:%Y%m%dT%H%M%S}_{creation_time:%Y%m%dT%H%M%S}_{duration:4d}_{cycle:3d}_{relative_orbit:3d}_{frame:4d}_{centre:3s}_{mode:1s}_{timeliness:2s}_{collection:3s}.SEN3/geo_coordinates.nc']
